Clean command bar and shared UI plumbing
