+++ scripts/build-electrobun-view.ts
@@ -1,93 +1,27 @@
-import { mkdir, readFile, rm, writeFile } from "fs/promises";
-import { join, relative } from "path";
-import { TITLEBAR_OVERLAY_HEIGHT_PX } from "../src/components/layout/titlebar-overlay";
+import { mkdir, rm } from "fs/promises";
+import { join } from "path";
+import {
+  electrobunViewPath,
+  writeElectrobunViewPage,
+} from "../src/renderers/electrobun/view/build-assets";
 
 const outdir = join(process.cwd(), "dist", "electrobun-view");
-const electrobunViewDir = join(process.cwd(), "src", "renderers", "electrobun", "view");
-
-interface ResolveArgs {
-  path: string;
-  importer?: string;
-}
-
-function aliasImport(args: ResolveArgs, sourceSuffix: string, target: string) {
-  if (args.path.endsWith(sourceSuffix)) {
-    return { path: join(electrobunViewDir, target) };
-  }
-  return undefined;
-}
-
-function aliasRelativeImport(args: ResolveArgs, importerSuffix: string, sourcePath: string, target: string) {
-  if (args.path === sourcePath && args.importer?.endsWith(importerSuffix)) {
-    return { path: join(electrobunViewDir, target) };
-  }
-  return undefined;
-}
 
 await rm(outdir, { recursive: true, force: true });
 await mkdir(outdir, { recursive: true });
 
-const result = await Bun.build({
-  entrypoints: [join(process.cwd(), "src", "renderers", "electrobun", "view", "main.tsx")],
+await writeElectrobunViewPage({
+  entrypoint: electrobunViewPath("main.tsx"),
   outdir,
-  target: "browser",
-  format: "esm",
-  splitting: false,
-  sourcemap: "external",
-  minify: true,
-  define: {
-    "process.env.NODE_ENV": "\"production\"",
-  },
-  plugins: [
-    {
-      name: "electrobun-renderer-native-bridges",
-      setup(build) {
-        build.onResolve({ filter: /.*/ }, (args) => (
-          aliasImport(args, "plugins/builtin/notes-files", "notes-files.ts")
-          ?? aliasImport(args, "notes-files", "notes-files.ts")
-          ?? aliasImport(args, "core/app-services", "app-services.ts")
-          ?? aliasImport(args, "kitty-support", "native-stubs/chart-kitty-support.ts")
-          ?? aliasImport(args, "native/kitty-support", "native-stubs/chart-kitty-support.ts")
-          ?? aliasRelativeImport(args, "components/chart/native/renderer-selection.ts", "./kitty-support", "native-stubs/chart-kitty-support.ts")
-          ?? aliasImport(args, "native/surface-manager", "native-stubs/chart-surface-manager.ts")
-          ?? aliasImport(args, "native/surface-sync", "native-stubs/chart-surface-sync.ts")
-        ));
-      },
-    },
+  pluginName: "electrobun-renderer-native-bridges",
+  extraAliasRules: [
+    ["core/app-services", "app-services.ts"],
   ],
-});
-
-if (!result.success) {
-  for (const log of result.logs) {
-    console.error(log);
-  }
-  process.exitCode = 1;
-  throw new Error("Failed to build Electrobun view assets");
-}
-
-const entry = result.outputs.find((output) => output.kind === "entry-point" && output.path.endsWith(".js"));
-if (!entry) {
-  throw new Error("Electrobun view build did not produce a JavaScript entrypoint");
-}
-
-const entrySrc = `./${relative(outdir, entry.path).replaceAll("\\", "/")}`;
-const stylesheet = (await readFile(join(electrobunViewDir, "styles.css"), "utf8"))
-  .replaceAll("__TITLEBAR_OVERLAY_HEIGHT_PX__", String(TITLEBAR_OVERLAY_HEIGHT_PX));
-
-await writeFile(join(outdir, "index.html"), renderElectrobunViewHtml(entrySrc, stylesheet));
-
-function renderElectrobunViewHtml(entrySrc: string, stylesheet: string): string {
-  return `<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Gloomberb</title>
-    <style>${stylesheet}</style>
-  </head>
-  <body style="margin:0;background:#000;">
-    <div id="root"><div class="gloom-loading">Loading Gloomberb...</div></div>
-    <script>
+  failureMessage: "Failed to build Electrobun view assets",
+  missingEntryMessage: "Electrobun view build did not produce a JavaScript entrypoint",
+  title: "Gloomberb",
+  loadingText: "Loading Gloomberb...",
+  bootstrapScript: `
       const bootstrapFatalHtml = [
         '<div class="gloom-fatal">',
         '<h1>Gloomberb failed to start</h1>',
@@ -129,9 +63,5 @@ function renderElectrobunViewHtml(entrySrc: string, stylesheet: string): string 
       ));
       window.addEventListener("unhandledrejection", (event) => renderBootstrapError(event.reason, "", "unhandledrejection"));
       document.getElementById("root").innerHTML = '<div class="gloom-loading">Booting Gloomberb renderer...</div>';
-    </script>
-    <script type="module" src="${entrySrc}"></script>
-  </body>
-</html>
-`;
-}
+`,
+});
