+++ src/components/layout/pane-sizing.ts
@@ -12,23 +12,55 @@ export function shouldReservePaneFooter(nativePaneChrome: boolean | undefined, s
   return !nativePaneChrome || showFooter;
 }
 
-export function getPaneBodyHeight(height: number, reserveFooter = true): number {
-  const cellHeight = Math.max(1, Math.floor(height));
-  return Math.max(1, cellHeight - PANE_HEADER_ROWS - (reserveFooter ? PANE_FOOTER_ROWS : 0));
+export function resolvePaneBodyHeight({
+  height,
+  nativePaneChrome,
+  reserveFooter = true,
+  headerRows = PANE_HEADER_ROWS,
+}: {
+  height: number;
+  nativePaneChrome?: boolean;
+  reserveFooter?: boolean;
+  headerRows?: number;
+}): number {
+  const finiteHeight = Number.isFinite(height) ? height : 1;
+  const normalizedHeight = nativePaneChrome ? finiteHeight : Math.max(1, Math.floor(finiteHeight));
+  return Math.max(1, normalizedHeight - headerRows - (reserveFooter ? PANE_FOOTER_ROWS : 0));
 }
 
-export function getPaneBodyWidth(width: number): number {
-  return Math.max(1, Math.floor(width));
+export function getPaneBodyLayoutProps(nativePaneChrome: boolean | undefined, bodyHeight: number | undefined) {
+  if (nativePaneChrome) return NATIVE_PANE_BODY_LAYOUT_PROPS;
+  return {
+    height: bodyHeight,
+    flexGrow: bodyHeight == null ? 1 : 0,
+    flexBasis: bodyHeight == null ? 0 : undefined,
+  };
 }
 
-export function getNativePaneBodyHeight(height: number, reserveFooter = true): number {
-  const preciseHeight = Number.isFinite(height) ? height : 1;
-  return Math.max(
-    1,
-    preciseHeight - PANE_HEADER_ROWS - (reserveFooter ? PANE_FOOTER_ROWS : 0),
-  );
+export function resolvePaneBodyFrame({
+  width,
+  height,
+  nativePaneChrome,
+  reserveFooter = true,
+  headerRows = PANE_HEADER_ROWS,
+}: {
+  width?: number;
+  height?: number;
+  nativePaneChrome?: boolean;
+  reserveFooter?: boolean;
+  headerRows?: number;
+}) {
+  const bodyHeight = typeof height === "number"
+    ? resolvePaneBodyHeight({ height, nativePaneChrome, reserveFooter, headerRows })
+    : undefined;
+  return {
+    width: typeof width === "number" ? resolvePaneBodyWidth(width, nativePaneChrome) : undefined,
+    height: bodyHeight,
+    layoutProps: getPaneBodyLayoutProps(nativePaneChrome, bodyHeight),
+  };
 }
 
-export function getNativePaneBodyWidth(width: number): number {
-  return Math.max(1, Number.isFinite(width) ? width : 1);
+function resolvePaneBodyWidth(width: number, nativePaneChrome: boolean | undefined): number {
+  const finiteWidth = Number.isFinite(width) ? width : 1;
+  return nativePaneChrome ? Math.max(1, finiteWidth) : Math.max(1, Math.floor(finiteWidth));
 }
