+++ src/components/layout/status-bar.test.tsx
@@ -45,37 +45,6 @@ describe("StatusBar", () => {
     expect(actions).toContainEqual({ type: "SET_COMMAND_BAR", open: true, query: "" });
   });
 
-  test("renders layout tabs without preview suffixes", async () => {
-    const config = createDefaultConfig("/tmp/gloomberb-test");
-    const researchLayout = cloneLayout(config.layout);
-    researchLayout.dockRoot = { kind: "pane", instanceId: "portfolio-list:main" };
-    researchLayout.floating = [{ instanceId: "ticker-detail:main", x: 8, y: 2, width: 36, height: 12 }];
-
-    const state = {
-      ...createInitialState({
-        ...config,
-        layouts: [
-          { name: "Default", layout: cloneLayout(config.layout) },
-          { name: "Research", layout: researchLayout },
-        ],
-      }),
-      statusBarVisible: true,
-    };
-
-    testSetup = await testRender(
-      <AppContext value={{ state, dispatch: () => {} }}>
-        <StatusBar />
-      </AppContext>,
-      { width: 120, height: 1 },
-    );
-
-    await testSetup.renderOnce();
-
-    const frame = testSetup.captureCharFrame();
-    expect(frame).toContain("Default");
-    expect(frame).toContain("Research");
-  });
-
   test("shows a gridlock tip after a corner snap and runs gridlock on click", async () => {
     const config = createDefaultConfig("/tmp/gloomberb-test");
     const floatingLayout = cloneLayout(config.layout);
@@ -185,74 +154,5 @@ describe("StatusBar", () => {
       globalThis.setTimeout = originalSetTimeout;
       globalThis.clearTimeout = originalClearTimeout;
     }
-  });
-
-  test("omits focused ticker quote details", async () => {
-    const config = createDefaultConfig("/tmp/gloomberb-test");
-    config.layout.instances = config.layout.instances.map((instance) => (
-      instance.instanceId === "ticker-detail:main"
-        ? { ...instance, binding: { kind: "fixed" as const, symbol: "AMD" } }
-        : instance
-    ));
-    const state = createInitialState(config);
-    state.statusBarVisible = true;
-    state.focusedPaneId = "ticker-detail:main";
-    state.tickers = new Map([["AMD", {
-      metadata: {
-        ticker: "AMD",
-        exchange: "NASDAQ",
-        currency: "USD",
-        name: "Advanced Micro Devices",
-        portfolios: [],
-        watchlists: [],
-        positions: [],
-        custom: {},
-        tags: [],
-      },
-    }]]);
-    state.financials = new Map([["AMD", {
-      annualStatements: [],
-      quarterlyStatements: [],
-      priceHistory: [],
-      quote: {
-        symbol: "AMD",
-        providerId: "ibkr",
-        dataSource: "live",
-        price: 100,
-        currency: "USD",
-        change: 1,
-        changePercent: 1,
-        lastUpdated: Date.now(),
-        listingExchangeName: "NASDAQ",
-        routingExchangeName: "SMART",
-        marketState: "PRE",
-        sessionConfidence: "derived",
-        preMarketPrice: 101,
-        preMarketChange: 2,
-        preMarketChangePercent: 2,
-        provenance: {
-          price: { providerId: "ibkr", dataSource: "live" },
-          session: { providerId: "yahoo", dataSource: "delayed" },
-        },
-      },
-    }]]);
-
-    testSetup = await testRender(
-      <AppContext value={{ state, dispatch: () => {} }}>
-        <StatusBar />
-      </AppContext>,
-      { width: 120, height: 1 },
-    );
-
-    await testSetup.renderOnce();
-
-    const frame = testSetup.captureCharFrame();
-    expect(frame).not.toContain("AMD");
-    expect(frame).not.toContain("Pre");
-    expect(frame).not.toContain("101");
-    expect(frame).not.toContain("NASDAQ PRE-MKT");
-    expect(frame).not.toContain("px IBKR live");
-    expect(frame).not.toContain("ses Yahoo");
-    expect(frame).not.toContain("SMART OPEN");
   });
 });
